backend/requirements.txt
# Use Python 3.11 or 3.12 if pip install fails (e.g. Rust/build errors on 3.13)
fastapi>=0.104.1,<0.115
uvicorn[standard]>=0.24.0,<0.32
sqlalchemy>=2.0.23,<2.1
alembic>=1.12.1,<1.14
pydantic>=2.10.0,<3
pydantic-settings>=2.6.0,<3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.1,<5
python-multipart>=0.0.6
httpx>=0.25.2,<0.28
redis>=5.0.1
prometheus-client>=0.19.0
python-dotenv>=1.0.0
